Studio 2

+++ src/exercises/RectangleArea.java
@@ -5,9 +5,9 @@ public class RectangleArea {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("What is the length of the rectangle?");
-        int length = input.nextInt();
+        double length = input.nextDouble();
         System.out.println("What is the width of the rectangle?");
-        int width = input.nextInt();
+        double width = input.nextDouble();
         double area = length * width;
         System.out.println("The area of the rectangle is " + area);
     }
